Added permissions and restrictions to users CRUD

+++ app/resources/users.py
@@ -5,6 +5,7 @@ from flask_restful import Resource
 from marshmallow import ValidationError
 from flask_login import current_user, login_required
 
+from app.admin_required import admin_required
 from app.models import User, db
 from app.schemas import UserSchema
 
@@ -16,6 +17,7 @@ class UserListResource(Resource):
 
     @staticmethod
     @login_required
+    @admin_required
     def get():
         """Get users"""
         users = db.session.query(User).all()
@@ -41,37 +43,42 @@ class UserResource(Resource):
     @login_required
     def get(user_id):
         """Get user by id"""
-        user = User.query.get_or_404(user_id)
-        return user_schema.dump(user)
+        if current_user.user_id == user_id or current_user.is_admin:
+            user = User.query.get_or_404(user_id)
+            return user_schema.dump(user)
+        return {"Error": "You don't have permission to do that"}, 403
 
     @staticmethod
     @login_required
     def put(user_id):
         """Update a user"""
+        if current_user.user_id == user_id or current_user.is_admin:
+            user = db.session.query(User).filter_by(user_id=user_id).first()
+            if not user:
+                return {"Error": "User was not found"}, 404
 
-        user = db.session.query(User).filter_by(user_id=user_id).first()
-        if not user:
-            return {"Error": "User was not found"}, 404
+            try:
+                user = user_schema.load(
+                    request.json, instance=user, session=db.session
+                )
+            except ValidationError as error:
+                return {"Error": str(error)}, 400
 
-        try:
-            user = user_schema.load(
-                request.json, instance=user, session=db.session
-            )
-        except ValidationError as error:
-            return {"Error": str(error)}, 400
-
-        db.session.add(user)
-        db.session.commit()
-        return user_schema.dump(user), 200
+            db.session.add(user)
+            db.session.commit()
+            return user_schema.dump(user), 200
+        return {"Error": "You don't have permission to do that"}, 403
 
     @staticmethod
     @login_required
     def delete(user_id):
         """Delete user by id"""
-        user = User.query.get_or_404(user_id)
-        db.session.delete(user)
-        db.session.commit()
-        return jsonify({
-            "status": 200,
-            "reason": "User is deleted"
-        })
+        if current_user.user_id == user_id or current_user.is_admin:
+            user = User.query.get_or_404(user_id)
+            db.session.delete(user)
+            db.session.commit()
+            return jsonify({
+                "status": 200,
+                "reason": "User is deleted"
+            })
+        return {"Error": "You don't have permission to do that"}, 403
